requirements.txt
certifi >= 14.05.14
six >= 1.10
python_dateutil >= 2.5.3
setuptools >= 21.0.0
tzdata >= 2024.1
urllib3 >= 1.15.1
pyyaml
pymongo[srv]>=4.0.0
dnspython>=2.0.0
flask>=2.3.0
flask-cors>=4.0.0